generelt

+++ function/task5.js
@@ -3,6 +3,6 @@ var double = function (number) {
     return number * 2;
 };
 
-console.log(double(3));
+console.log(double(8));
 console.log(double(3) + double(10));
 console.log(double(double(6)));
